requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.5.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
chromadb==1.1.1
click==8.1.8
coloredlogs==15.0.1
crewai==1.9.3
crewai-tools==1.9.3
cryptography==47.0.0
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.18.0
durationpy==0.10
et_xmlfile==2.0.0
fastuuid==0.14.0
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.4.0
gitdb==4.0.12
GitPython==3.1.49
google_search_results==2.4.2
googleapis-common-protos==1.74.0
groq==0.37.1
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
identify==2.6.19
idna==3.13
importlib_metadata==8.5.0
importlib_resources==7.1.0
instructor==1.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lancedb==0.5.7
langchain-anthropic==1.4.2
langchain-core==1.3.2
langchain-groq==1.1.2
langchain-protocol==0.0.14
langsmith==0.7.38
litellm==1.83.14
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
narwhals==2.20.0
nodeenv==1.10.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.8
overrides==7.7.0
packaging==26.2
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
platformdirs==4.9.6
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.6.0
propcache==0.4.1
protobuf==5.29.6
py==1.11.0
pyarrow==24.0.0
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.2
Pygments==2.20.0
PyJWT==2.12.1
pylance==0.9.18
PyMuPDF==1.26.7
PyPDF2==3.0.1
pypdfium2==5.7.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.27
pytube==15.0.0
PyYAML==6.0.3
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.3.4
rpds-py==0.30.0
semver==3.0.4
serpapi==1.0.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.4.1
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.9.30
uvicorn==0.46.0
uvloop==0.22.1
virtualenv==21.3.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
youtube-transcript-api==1.2.4
zipp==3.23.1
zstandard==0.25.0
